Decouple Mcp handling from transport
